requirements.txt
# ─────────────────────────────────────────
#  EduCenter LMS — Django Backend
#  Requirements v1.0.0
# ─────────────────────────────────────────

# Core
Django==5.0.6
djangorestframework==3.15.2

# Authentication (JWT)
djangorestframework-simplejwt==5.3.1

# CORS (React bilan ishlash uchun)
django-cors-headers==4.4.0

# Filtering & Search
django-filter==24.2

# API Documentation (Swagger / ReDoc)
drf-spectacular==0.27.2

# Environment variables (.env fayl)
python-decouple==3.8

# Image upload
Pillow==10.4.0

# File & Media Storage
django-storages==1.14.4
boto3==1.34.144                # AWS S3/MinIO support

# Database
psycopg2-binary==2.9.9        # PostgreSQL uchun

# Validation (extra)
drf-writable-nested==0.7.0    # Nested serializer CRUD

# Pagination & Utils
djangorestframework-camel-case==1.4.2  # camelCase JSON (React uchun qulay)

# Development tools
black==24.8.0                 # Code formatter
flake8==7.1.1                 # Linter
isort==5.13.2                 # Import sorter

# Testing
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.1            # Test uchun fake data

# Production
gunicorn==22.0.0              # WSGI server
whitenoise==6.7.0             # Static files serving